requirements.txt
Flask
gunicorn
pandas
openpyxl
qrcode
pillow
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
